Filter Ions by Group

Filter Ions by Group

Usage

filter_group(
  mpactr_object,
  group_threshold = 0.01,
  group_to_remove,
  remove_ions = TRUE,
  copy_object = FALSE
)

Arguments

mpactr_object

An mpactr_object. See import_data().

group_threshold

Relative abundance threshold at which to remove ions. Default = 0.01.

group_to_remove

Biological group name to remove ions from.

remove_ions

A logical. If TRUE failing ions will be removed from the peak table. Default = TRUE.

copy_object

A boolean paramter that allows users to return a copied object instead of modifying the object.

Value

an mpactr_object

Details

filter_group() removes ions which have a signigifanct presence in the defined group (typically solvent blanks). Significance is determined by relative ion abundance arcoss groups at the user-defined group_threshold.

Filtering is assessed on function parameter group_threshold. The defualt is 0.01, meaning features present in group_to_remove at greater than 1% are removed.

Examples

data <- import_data(example("coculture_peak_table.csv"),
                    example("metadata.csv"),
                    format = "Progenesis")

data_filter <- filter_group(data,
                              group_threshold = 0.01,
                              group_to_remove = "Blanks",
                              remove_ions = TRUE)
#> ℹ Parsing 1303 peaks based on the following sample group: Blanks.
#> ℹ Argument remove_ions is: TRUE. Removing peaks from Blanks.
#> ✔ 796 ions failed the Blanks filter, 507 ions remain.