Return the summary for a single mpactR filter.

filter_summary() is a wrapper function to return the summary from a single filter within the given mpactr object.

Usage

filter_summary(mpactr_object, filter, group = NULL)

Arguments

mpactr_object

The mpactr object that is created by calling the import_data() function.

filter

The name of a filter whose summary is to be extracted. Must be one of: "mispicked", "group", "replicability", or "insource".

group

If filter = "group", the name of the Biological_Group used to filter.

Value

a list reporting 1) compound ids for compounds which failed the filter and 2) compound ids for compounds which passed the filter.

Examples

data <- import_data(example("coculture_peak_table.csv"),
                    example("metadata.csv"),
                    format = "Progenesis")

data_filter <- filter_mispicked_ions(data)
#> ℹ Checking 1303 peaks for mispicked peaks.
#> ℹ Argument merge_peaks is: TRUE. Merging mispicked peaks with method sum.
#> ✔ 70 ions failed the mispicked filter, 1233 ions remain.

mispicked_summary <- filter_summary(data_filter, filter = "mispicked")
mispicked_summary
#> $failed_ions
#>  [1] 1189  945  896 1210 1249  884 1060 1008 1304 1271 1333  937 1196 1014  895
#> [16] 1228  817  853 1122  938 1145 1021  946  161 1032 1033  282  154  779 1254
#> [31] 1224  260  264   18 1012  800 1251  901  969 1259 1223  528 1095  927 1242
#> [46]  410 1260  427 1157 1206  393 1252 1204  975  584 1282 1090 1257  507 1087
#> [61] 1153 1261  371  373  340 1267  546  495 1289 1290
#> 
#> $passed_ions
#>    [1]    1    2    3    4    5    6    7    8    9   10   11   12   13   14
#>   [15]   15   16   17   19   20   21   22   23   24   25   26   27   28   29
#>   [29]   30   31   32   33   34   35   36   37   38   39   40   41   42   43
#>   [43]   44   45   46   47   48   49   50   51   52   53   54   55   56   57
#>   [57]   58   59   60   61   62   63   64   65   66   67   68   69   70   71
#>   [71]   72   73   74   75   76   77   78   79   80   81   82   83   84   85
#>   [85]   86   87   88   89   90   91   92   93   94   95   96   97   98   99
#>   [99]  100  101  102  103  104  105  106  107  108  109  110  111  112  113
#>  [113]  114  115  116  117  118  119  120  121  122  123  124  125  126  127
#>  [127]  128  129  130  131  132  133  134  135  136  137  138  139  140  141
#>  [141]  142  143  144  145  146  147  148  149  150  151  152  153  155  156
#>  [155]  157  158  159  160  162  163  164  165  166  167  168  169  170  171
#>  [169]  172  173  174  175  176  177  178  179  180  181  182  183  184  185
#>  [183]  186  187  188  189  190  191  192  193  194  195  196  197  198  199
#>  [197]  200  201  202  203  204  205  206  207  208  209  210  211  212  213
#>  [211]  214  215  216  217  218  219  220  221  222  223  224  225  226  227
#>  [225]  228  229  230  231  232  233  234  235  236  237  238  239  240  241
#>  [239]  242  243  244  245  246  247  248  249  250  251  252  253  254  255
#>  [253]  256  257  258  259  261  262  263  265  266  267  268  269  270  271
#>  [267]  272  273  274  275  276  277  278  279  280  281  283  284  285  286
#>  [281]  287  288  289  290  291  292  293  294  295  296  297  298  299  300
#>  [295]  301  302  303  304  305  306  307  308  309  310  311  312  313  314
#>  [309]  315  316  317  318  319  320  321  322  323  324  325  326  327  328
#>  [323]  329  330  331  332  333  334  335  336  337  338  339  341  342  343
#>  [337]  344  345  346  347  348  349  350  351  352  353  354  355  356  357
#>  [351]  358  359  360  361  362  363  364  365  366  367  368  369  370  372
#>  [365]  374  375  376  377  378  379  380  381  382  383  384  385  386  387
#>  [379]  388  389  390  391  392  394  395  396  397  398  399  400  401  402
#>  [393]  403  404  405  406  407  408  409  411  412  413  414  415  416  417
#>  [407]  418  419  420  421  422  423  424  425  426  428  429  430  431  432
#>  [421]  433  434  435  436  437  438  439  440  441  442  443  444  445  446
#>  [435]  447  448  449  450  451  452  453  454  455  456  457  458  459  460
#>  [449]  461  462  463  464  465  466  467  468  469  470  471  472  473  474
#>  [463]  475  476  477  478  479  480  481  482  483  484  485  486  487  488
#>  [477]  489  490  491  492  493  494  496  497  498  499  500  501  502  503
#>  [491]  504  505  506  508  509  510  511  512  513  514  515  516  517  518
#>  [505]  519  520  521  522  523  524  525  526  527  529  530  531  532  533
#>  [519]  534  535  536  537  538  539  540  541  542  543  544  545  547  548
#>  [533]  549  550  551  552  553  554  555  556  557  558  559  560  561  562
#>  [547]  563  564  565  566  567  568  569  570  571  572  573  574  575  576
#>  [561]  577  578  579  580  581  582  583  585  586  587  588  589  590  591
#>  [575]  592  593  594  595  596  597  598  599  600  601  602  603  604  605
#>  [589]  606  607  608  609  610  611  612  613  614  615  616  617  618  619
#>  [603]  620  621  622  623  624  625  626  627  628  629  630  631  632  633
#>  [617]  634  635  636  637  638  639  640  641  642  643  644  645  646  647
#>  [631]  648  649  650  651  652  653  655  656  657  658  659  660  661  662
#>  [645]  663  664  665  666  667  668  669  670  671  672  673  674  675  676
#>  [659]  677  678  679  680  681  682  683  685  686  687  688  689  690  691
#>  [673]  692  693  694  695  696  699  700  701  702  703  704  705  706  707
#>  [687]  708  710  711  712  715  716  718  719  721  722  723  724  726  727
#>  [701]  728  729  732  733  734  738  739  741  742  744  745  746  748  749
#>  [715]  751  753  758  759  760  761  762  763  764  765  766  767  768  769
#>  [729]  770  771  772  773  774  775  776  777  778  780  781  782  783  784
#>  [743]  785  786  787  788  789  790  791  792  793  794  795  796  797  798
#>  [757]  799  801  802  803  804  805  806  807  808  809  810  811  812  813
#>  [771]  814  815  816  818  819  820  821  822  823  824  825  826  827  828
#>  [785]  829  830  831  832  833  834  835  836  837  838  839  840  841  842
#>  [799]  843  844  845  846  847  848  849  850  851  852  854  855  856  857
#>  [813]  858  859  860  861  862  863  864  865  866  867  868  869  870  871
#>  [827]  872  873  874  875  876  877  878  879  880  881  882  883  885  886
#>  [841]  887  888  889  890  891  892  893  894  897  898  899  900  902  903
#>  [855]  904  905  906  907  908  909  910  911  912  913  914  915  916  917
#>  [869]  918  919  920  921  922  923  924  925  926  928  929  930  931  932
#>  [883]  933  934  935  936  939  940  941  942  943  944  947  948  949  950
#>  [897]  951  952  953  954  955  956  957  958  959  960  961  962  963  964
#>  [911]  965  966  967  968  970  971  972  973  974  976  977  978  979  980
#>  [925]  981  982  983  984  985  986  987  988  989  990  991  992  993  994
#>  [939]  995  996  997  998  999 1000 1001 1002 1003 1004 1005 1006 1007 1009
#>  [953] 1010 1011 1013 1015 1016 1017 1018 1019 1020 1022 1023 1024 1025 1026
#>  [967] 1027 1028 1029 1030 1031 1034 1035 1036 1037 1038 1039 1040 1041 1042
#>  [981] 1043 1044 1045 1046 1047 1048 1050 1051 1052 1053 1054 1055 1056 1057
#>  [995] 1058 1059 1061 1062 1063 1064 1065 1066 1067 1068 1069 1070 1071 1072
#> [1009] 1073 1074 1075 1076 1077 1078 1079 1080 1081 1082 1083 1084 1085 1086
#> [1023] 1088 1089 1091 1092 1093 1094 1096 1097 1098 1099 1100 1101 1102 1103
#> [1037] 1104 1105 1106 1107 1108 1109 1110 1111 1112 1113 1114 1115 1116 1117
#> [1051] 1118 1119 1120 1121 1123 1124 1125 1126 1127 1128 1129 1130 1131 1132
#> [1065] 1133 1134 1135 1136 1137 1138 1139 1140 1141 1142 1143 1144 1146 1147
#> [1079] 1148 1149 1150 1151 1152 1154 1155 1156 1158 1159 1160 1161 1162 1163
#> [1093] 1164 1165 1166 1167 1168 1169 1170 1171 1172 1173 1174 1175 1176 1177
#> [1107] 1178 1179 1180 1181 1182 1183 1184 1185 1186 1187 1188 1190 1191 1192
#> [1121] 1193 1194 1195 1197 1198 1199 1200 1201 1202 1203 1205 1207 1208 1209
#> [1135] 1211 1212 1213 1214 1215 1216 1217 1218 1219 1220 1221 1222 1225 1226
#> [1149] 1227 1229 1230 1231 1232 1233 1234 1235 1236 1237 1238 1239 1240 1241
#> [1163] 1243 1244 1245 1246 1247 1250 1253 1255 1256 1258 1262 1263 1264 1265
#> [1177] 1266 1268 1269 1270 1272 1273 1274 1276 1277 1278 1279 1280 1281 1283
#> [1191] 1284 1285 1286 1287 1288 1291 1292 1293 1294 1295 1296 1297 1298 1299
#> [1205] 1300 1301 1302 1303 1305 1306 1307 1308 1309 1310 1311 1312 1313 1314
#> [1219] 1315 1316 1317 1318 1319 1320 1321 1322 1323 1327 1328 1329 1330 1332
#> [1233] 1334
#>