mpactr: Correction of Preprocessed MS Data

An 'R' implementation of the 'python' program Metabolomics Peak Analysis Computational Tool ('MPACT') (Robert M. Samples, Sara P. Puckett, and Marcy J. Balunas (2023) doi:10.1021/acs.analchem.2c04632 ). Filters in the package serve to address common errors in tandem mass spectrometry preprocessing, including: (1) isotopic patterns that are incorrectly split during preprocessing, (2) features present in solvent blanks due to carryover between samples, (3) features whose abundance is greater than user-defined abundance threshold in a specific group of samples, for example media blanks, (4) ions that are inconsistent between technical replicates, and (5) in-source fragment ions created during ionization before fragmentation in the tandem mass spectrometry workflow.

See also

Author

Maintainer: Patrick Schloss pschloss@umich.edu (ORCID) [copyright holder]

Authors:

Allison Mason masonar@umich.edu (ORCID)

Gregory Johnson grejoh@umich.edu