Return the summary for a single mpactr filter.

filter_summary() is a wrapper function to return the summary from a single filter within the given mpactr object.

Usage

filter_summary(mpactr_object, filter, group = NULL)

Arguments

mpactr_object

The mpactr object that is created by calling the import_data() function.

filter

The name of a filter whose summary is to be extracted. Must be one of: "mispicked", "group", "replicability", or "insource".

group

If filter = "group", the name of the Biological_Group used to filter.

Value

a list reporting 1) compound ids for compounds which failed the filter and 2) compound ids for compounds which passed the filter.

Examples

data <- import_data(example_path("coculture_peak_table.csv"),
  example_path("metadata.csv"),
  format = "Progenesis"
)

data_filter <- filter_mispicked_ions(data)
#> ℹ Checking 1303 peaks for mispicked peaks.
#> ℹ Argument merge_peaks is: TRUE. Merging mispicked peaks with method sum.
#> ✔ 70 ions failed the mispicked filter, 1233 ions remain.

mispicked_summary <- filter_summary(data_filter, filter = "mispicked")
mispicked_summary
#> $failed_ions
#>  [1] "1189" "945"  "896"  "1210" "1249" "884"  "1060" "1008" "1304" "1271"
#> [11] "1333" "937"  "1196" "1014" "895"  "1228" "817"  "853"  "1122" "938" 
#> [21] "1145" "1021" "946"  "161"  "1032" "1033" "282"  "154"  "779"  "1254"
#> [31] "1224" "260"  "264"  "18"   "1012" "800"  "1251" "901"  "969"  "1259"
#> [41] "1223" "528"  "1095" "927"  "1242" "410"  "1260" "427"  "1157" "1206"
#> [51] "393"  "1252" "1204" "975"  "584"  "1282" "1090" "1257" "507"  "1087"
#> [61] "1153" "1261" "371"  "373"  "340"  "1267" "546"  "495"  "1289" "1290"
#> 
#> $passed_ions
#>    [1] "1"    "10"   "100"  "1000" "1001" "1002" "1003" "1004" "1005" "1006"
#>   [11] "1007" "1009" "101"  "1010" "1011" "1013" "1015" "1016" "1017" "1018"
#>   [21] "1019" "102"  "1020" "1022" "1023" "1024" "1025" "1026" "1027" "1028"
#>   [31] "1029" "103"  "1030" "1031" "1034" "1035" "1036" "1037" "1038" "1039"
#>   [41] "104"  "1040" "1041" "1042" "1043" "1044" "1045" "1046" "1047" "1048"
#>   [51] "105"  "1050" "1051" "1052" "1053" "1054" "1055" "1056" "1057" "1058"
#>   [61] "1059" "106"  "1061" "1062" "1063" "1064" "1065" "1066" "1067" "1068"
#>   [71] "1069" "107"  "1070" "1071" "1072" "1073" "1074" "1075" "1076" "1077"
#>   [81] "1078" "1079" "108"  "1080" "1081" "1082" "1083" "1084" "1085" "1086"
#>   [91] "1088" "1089" "109"  "1091" "1092" "1093" "1094" "1096" "1097" "1098"
#>  [101] "1099" "11"   "110"  "1100" "1101" "1102" "1103" "1104" "1105" "1106"
#>  [111] "1107" "1108" "1109" "111"  "1110" "1111" "1112" "1113" "1114" "1115"
#>  [121] "1116" "1117" "1118" "1119" "112"  "1120" "1121" "1123" "1124" "1125"
#>  [131] "1126" "1127" "1128" "1129" "113"  "1130" "1131" "1132" "1133" "1134"
#>  [141] "1135" "1136" "1137" "1138" "1139" "114"  "1140" "1141" "1142" "1143"
#>  [151] "1144" "1146" "1147" "1148" "1149" "115"  "1150" "1151" "1152" "1154"
#>  [161] "1155" "1156" "1158" "1159" "116"  "1160" "1161" "1162" "1163" "1164"
#>  [171] "1165" "1166" "1167" "1168" "1169" "117"  "1170" "1171" "1172" "1173"
#>  [181] "1174" "1175" "1176" "1177" "1178" "1179" "118"  "1180" "1181" "1182"
#>  [191] "1183" "1184" "1185" "1186" "1187" "1188" "119"  "1190" "1191" "1192"
#>  [201] "1193" "1194" "1195" "1197" "1198" "1199" "12"   "120"  "1200" "1201"
#>  [211] "1202" "1203" "1205" "1207" "1208" "1209" "121"  "1211" "1212" "1213"
#>  [221] "1214" "1215" "1216" "1217" "1218" "1219" "122"  "1220" "1221" "1222"
#>  [231] "1225" "1226" "1227" "1229" "123"  "1230" "1231" "1232" "1233" "1234"
#>  [241] "1235" "1236" "1237" "1238" "1239" "124"  "1240" "1241" "1243" "1244"
#>  [251] "1245" "1246" "1247" "125"  "1250" "1253" "1255" "1256" "1258" "126" 
#>  [261] "1262" "1263" "1264" "1265" "1266" "1268" "1269" "127"  "1270" "1272"
#>  [271] "1273" "1274" "1276" "1277" "1278" "1279" "128"  "1280" "1281" "1283"
#>  [281] "1284" "1285" "1286" "1287" "1288" "129"  "1291" "1292" "1293" "1294"
#>  [291] "1295" "1296" "1297" "1298" "1299" "13"   "130"  "1300" "1301" "1302"
#>  [301] "1303" "1305" "1306" "1307" "1308" "1309" "131"  "1310" "1311" "1312"
#>  [311] "1313" "1314" "1315" "1316" "1317" "1318" "1319" "132"  "1320" "1321"
#>  [321] "1322" "1323" "1327" "1328" "1329" "133"  "1330" "1332" "1334" "134" 
#>  [331] "135"  "136"  "137"  "138"  "139"  "14"   "140"  "141"  "142"  "143" 
#>  [341] "144"  "145"  "146"  "147"  "148"  "149"  "15"   "150"  "151"  "152" 
#>  [351] "153"  "155"  "156"  "157"  "158"  "159"  "16"   "160"  "162"  "163" 
#>  [361] "164"  "165"  "166"  "167"  "168"  "169"  "17"   "170"  "171"  "172" 
#>  [371] "173"  "174"  "175"  "176"  "177"  "178"  "179"  "180"  "181"  "182" 
#>  [381] "183"  "184"  "185"  "186"  "187"  "188"  "189"  "19"   "190"  "191" 
#>  [391] "192"  "193"  "194"  "195"  "196"  "197"  "198"  "199"  "2"    "20"  
#>  [401] "200"  "201"  "202"  "203"  "204"  "205"  "206"  "207"  "208"  "209" 
#>  [411] "21"   "210"  "211"  "212"  "213"  "214"  "215"  "216"  "217"  "218" 
#>  [421] "219"  "22"   "220"  "221"  "222"  "223"  "224"  "225"  "226"  "227" 
#>  [431] "228"  "229"  "23"   "230"  "231"  "232"  "233"  "234"  "235"  "236" 
#>  [441] "237"  "238"  "239"  "24"   "240"  "241"  "242"  "243"  "244"  "245" 
#>  [451] "246"  "247"  "248"  "249"  "25"   "250"  "251"  "252"  "253"  "254" 
#>  [461] "255"  "256"  "257"  "258"  "259"  "26"   "261"  "262"  "263"  "265" 
#>  [471] "266"  "267"  "268"  "269"  "27"   "270"  "271"  "272"  "273"  "274" 
#>  [481] "275"  "276"  "277"  "278"  "279"  "28"   "280"  "281"  "283"  "284" 
#>  [491] "285"  "286"  "287"  "288"  "289"  "29"   "290"  "291"  "292"  "293" 
#>  [501] "294"  "295"  "296"  "297"  "298"  "299"  "3"    "30"   "300"  "301" 
#>  [511] "302"  "303"  "304"  "305"  "306"  "307"  "308"  "309"  "31"   "310" 
#>  [521] "311"  "312"  "313"  "314"  "315"  "316"  "317"  "318"  "319"  "32"  
#>  [531] "320"  "321"  "322"  "323"  "324"  "325"  "326"  "327"  "328"  "329" 
#>  [541] "33"   "330"  "331"  "332"  "333"  "334"  "335"  "336"  "337"  "338" 
#>  [551] "339"  "34"   "341"  "342"  "343"  "344"  "345"  "346"  "347"  "348" 
#>  [561] "349"  "35"   "350"  "351"  "352"  "353"  "354"  "355"  "356"  "357" 
#>  [571] "358"  "359"  "36"   "360"  "361"  "362"  "363"  "364"  "365"  "366" 
#>  [581] "367"  "368"  "369"  "37"   "370"  "372"  "374"  "375"  "376"  "377" 
#>  [591] "378"  "379"  "38"   "380"  "381"  "382"  "383"  "384"  "385"  "386" 
#>  [601] "387"  "388"  "389"  "39"   "390"  "391"  "392"  "394"  "395"  "396" 
#>  [611] "397"  "398"  "399"  "4"    "40"   "400"  "401"  "402"  "403"  "404" 
#>  [621] "405"  "406"  "407"  "408"  "409"  "41"   "411"  "412"  "413"  "414" 
#>  [631] "415"  "416"  "417"  "418"  "419"  "42"   "420"  "421"  "422"  "423" 
#>  [641] "424"  "425"  "426"  "428"  "429"  "43"   "430"  "431"  "432"  "433" 
#>  [651] "434"  "435"  "436"  "437"  "438"  "439"  "44"   "440"  "441"  "442" 
#>  [661] "443"  "444"  "445"  "446"  "447"  "448"  "449"  "45"   "450"  "451" 
#>  [671] "452"  "453"  "454"  "455"  "456"  "457"  "458"  "459"  "46"   "460" 
#>  [681] "461"  "462"  "463"  "464"  "465"  "466"  "467"  "468"  "469"  "47"  
#>  [691] "470"  "471"  "472"  "473"  "474"  "475"  "476"  "477"  "478"  "479" 
#>  [701] "48"   "480"  "481"  "482"  "483"  "484"  "485"  "486"  "487"  "488" 
#>  [711] "489"  "49"   "490"  "491"  "492"  "493"  "494"  "496"  "497"  "498" 
#>  [721] "499"  "5"    "50"   "500"  "501"  "502"  "503"  "504"  "505"  "506" 
#>  [731] "508"  "509"  "51"   "510"  "511"  "512"  "513"  "514"  "515"  "516" 
#>  [741] "517"  "518"  "519"  "52"   "520"  "521"  "522"  "523"  "524"  "525" 
#>  [751] "526"  "527"  "529"  "53"   "530"  "531"  "532"  "533"  "534"  "535" 
#>  [761] "536"  "537"  "538"  "539"  "54"   "540"  "541"  "542"  "543"  "544" 
#>  [771] "545"  "547"  "548"  "549"  "55"   "550"  "551"  "552"  "553"  "554" 
#>  [781] "555"  "556"  "557"  "558"  "559"  "56"   "560"  "561"  "562"  "563" 
#>  [791] "564"  "565"  "566"  "567"  "568"  "569"  "57"   "570"  "571"  "572" 
#>  [801] "573"  "574"  "575"  "576"  "577"  "578"  "579"  "58"   "580"  "581" 
#>  [811] "582"  "583"  "585"  "586"  "587"  "588"  "589"  "59"   "590"  "591" 
#>  [821] "592"  "593"  "594"  "595"  "596"  "597"  "598"  "599"  "6"    "60"  
#>  [831] "600"  "601"  "602"  "603"  "604"  "605"  "606"  "607"  "608"  "609" 
#>  [841] "61"   "610"  "611"  "612"  "613"  "614"  "615"  "616"  "617"  "618" 
#>  [851] "619"  "62"   "620"  "621"  "622"  "623"  "624"  "625"  "626"  "627" 
#>  [861] "628"  "629"  "63"   "630"  "631"  "632"  "633"  "634"  "635"  "636" 
#>  [871] "637"  "638"  "639"  "64"   "640"  "641"  "642"  "643"  "644"  "645" 
#>  [881] "646"  "647"  "648"  "649"  "65"   "650"  "651"  "652"  "653"  "655" 
#>  [891] "656"  "657"  "658"  "659"  "66"   "660"  "661"  "662"  "663"  "664" 
#>  [901] "665"  "666"  "667"  "668"  "669"  "67"   "670"  "671"  "672"  "673" 
#>  [911] "674"  "675"  "676"  "677"  "678"  "679"  "68"   "680"  "681"  "682" 
#>  [921] "683"  "685"  "686"  "687"  "688"  "689"  "69"   "690"  "691"  "692" 
#>  [931] "693"  "694"  "695"  "696"  "699"  "7"    "70"   "700"  "701"  "702" 
#>  [941] "703"  "704"  "705"  "706"  "707"  "708"  "71"   "710"  "711"  "712" 
#>  [951] "715"  "716"  "718"  "719"  "72"   "721"  "722"  "723"  "724"  "726" 
#>  [961] "727"  "728"  "729"  "73"   "732"  "733"  "734"  "738"  "739"  "74"  
#>  [971] "741"  "742"  "744"  "745"  "746"  "748"  "749"  "75"   "751"  "753" 
#>  [981] "758"  "759"  "76"   "760"  "761"  "762"  "763"  "764"  "765"  "766" 
#>  [991] "767"  "768"  "769"  "77"   "770"  "771"  "772"  "773"  "774"  "775" 
#> [1001] "776"  "777"  "778"  "78"   "780"  "781"  "782"  "783"  "784"  "785" 
#> [1011] "786"  "787"  "788"  "789"  "79"   "790"  "791"  "792"  "793"  "794" 
#> [1021] "795"  "796"  "797"  "798"  "799"  "8"    "80"   "801"  "802"  "803" 
#> [1031] "804"  "805"  "806"  "807"  "808"  "809"  "81"   "810"  "811"  "812" 
#> [1041] "813"  "814"  "815"  "816"  "818"  "819"  "82"   "820"  "821"  "822" 
#> [1051] "823"  "824"  "825"  "826"  "827"  "828"  "829"  "83"   "830"  "831" 
#> [1061] "832"  "833"  "834"  "835"  "836"  "837"  "838"  "839"  "84"   "840" 
#> [1071] "841"  "842"  "843"  "844"  "845"  "846"  "847"  "848"  "849"  "85"  
#> [1081] "850"  "851"  "852"  "854"  "855"  "856"  "857"  "858"  "859"  "86"  
#> [1091] "860"  "861"  "862"  "863"  "864"  "865"  "866"  "867"  "868"  "869" 
#> [1101] "87"   "870"  "871"  "872"  "873"  "874"  "875"  "876"  "877"  "878" 
#> [1111] "879"  "88"   "880"  "881"  "882"  "883"  "885"  "886"  "887"  "888" 
#> [1121] "889"  "89"   "890"  "891"  "892"  "893"  "894"  "897"  "898"  "899" 
#> [1131] "9"    "90"   "900"  "902"  "903"  "904"  "905"  "906"  "907"  "908" 
#> [1141] "909"  "91"   "910"  "911"  "912"  "913"  "914"  "915"  "916"  "917" 
#> [1151] "918"  "919"  "92"   "920"  "921"  "922"  "923"  "924"  "925"  "926" 
#> [1161] "928"  "929"  "93"   "930"  "931"  "932"  "933"  "934"  "935"  "936" 
#> [1171] "939"  "94"   "940"  "941"  "942"  "943"  "944"  "947"  "948"  "949" 
#> [1181] "95"   "950"  "951"  "952"  "953"  "954"  "955"  "956"  "957"  "958" 
#> [1191] "959"  "96"   "960"  "961"  "962"  "963"  "964"  "965"  "966"  "967" 
#> [1201] "968"  "97"   "970"  "971"  "972"  "973"  "974"  "976"  "977"  "978" 
#> [1211] "979"  "98"   "980"  "981"  "982"  "983"  "984"  "985"  "986"  "987" 
#> [1221] "988"  "989"  "99"   "990"  "991"  "992"  "993"  "994"  "995"  "996" 
#> [1231] "997"  "998"  "999" 
#>